requirements.txt
spacy
pymupdf
python-docx
streamlit
pandas
rapidfuzz
